Add some load timings to the report
Thought it might be useful for some to get a rough estimate of load times. (It is something I and my players notice, for sure.) For simplicity, I just added measurements from the init hook for this module to later hooks, such as ready and canvas ready. Pretty basics, really, but still possibly useful.

This of course, does not reflect the total time until Foundry is loaded. To get a better sense of the total load time, you would need to find a way to know when the user in fact started the log in. You could get closer by renaming this module to something like _debug-report, to get to the init hook faster, and marking it as a library module, which I think load first. Or maybe you or someone else knows how to get the load start time from Foundry API.